rename pins overview errors to service manager errors
