+++ admin/spec/lib/validator/pins-overview.validator.spec.js
@@ -1,7 +1,7 @@
 'use strict'
 /* global describe, it, expect */
 const pinsOverviewValidator = require('../../../lib/validator/pins-overview-validator')
-const pinsOverviewErrorMessages = require('../../../lib/errors/pins-overview')
+const serviceManagerErrorMessages = require('../../../lib/errors/service-manager')
 const checkWindowMock = require('../../mocks/check-window')
 const checkFormMock = require('../../mocks/check-form')
 
@@ -10,19 +10,19 @@ describe('pins-overview-validator', () => {
     const checkWindow = null
     const allForms = [checkFormMock]
     const error = pinsOverviewValidator.validate(checkWindow, allForms)
-    expect(error).toBe(pinsOverviewErrorMessages.noCurrentCheckWindow)
+    expect(error).toBe(serviceManagerErrorMessages.noCurrentCheckWindow)
   })
   it('returns an error if there are no check forms assigned', () => {
     const checkWindow = checkWindowMock
     const allForms = []
     const error = pinsOverviewValidator.validate(checkWindow, allForms)
-    expect(error).toBe(pinsOverviewErrorMessages.noCheckFormsAssigned)
+    expect(error).toBe(serviceManagerErrorMessages.noCheckFormsAssigned)
   })
   it('returns noCheckWindowAndForms error if there are no check forms assigned and there is no active checkwindow', () => {
     const checkWindow = null
     const allForms = []
     const error = pinsOverviewValidator.validate(checkWindow, allForms)
-    expect(error).toBe(pinsOverviewErrorMessages.noCheckWindowAndForms)
+    expect(error).toBe(serviceManagerErrorMessages.noCheckWindowAndForms)
   })
   it('returns an empty error object if there is an active check window and and check forms assigned', () => {
     const checkWindow = checkWindowMock
